refactor: rename currentSoundId to currentId for consistency in useHotkeys hook
